Tests: Shift tests to seekable stream supervisor location for coverage and review comments
